requirements.txt
pillow==11.2.1
torch==2.7.0
transformers==4.52.3
accelerate==1.7.0
openai==1.82.0
matplotlib==3.10.3
pycocoevalcap==1.2
torchmetrics==1.7.1